auto-wrap tab-completion in double quotes
when the string to be inserted includes a space character then the insert will get wrapped in double-quotes

this `insertString` function is _only_ called by tab-completion - it should probably be renamed to reflect that.  tab-completion looks for a “term” to complete.

as tab-completion completes the term the user has typed, it will keep what the user has already typed and just add to it (meaning that cases don’t change on user typed text)

when wrapping in double-quotes, the existing term to be completed will be moved right by 1 character and an opening double-quote inserted.  a closing double-quote gets added at the end.  the display is then updated to reflect the edit