autocomplete to omit hidden/system files
adds a `flags` byte to the `resolvePath` function, and in register C for the `mos_resolvepath` API, which is used to filter files based on their attributes

when the top-bit for the flags is set, then the lower bits specify all of the attributes that _must_ be present for a file/directory to be considered a match.  when the top bit is _not_ set the bits indicate attributes that should _not_ be present

the auto-complete system makes use of this, setting its resolvePath flags to `AM_HID | AM_SYS` - this ensures that hidden and system files will be omitted from its matches

currently all other uses of `resolvePath` use a flags byte of zero, ensuring all potential files will be matched.  in the future, some commands may change to make hidden/system files invisible to them by default, and gain flags to let them become visible