run files directly from the command line
`mos_exec`, the command interpreter function, will now fall back to attempting to use the `runfile` command with the input line if all other command matching attempts has failed. this means that typing a filename into the CLI will run it, if a runtype has been set up for that file type

also in this commit is a change to `mos_trim` so that by default it only removes non-space whitespace characters from the end of the passed in string.  this allows the `set` command to set variables that include spaces.

for compatibility, the `run` and `runbin` commands will trim spaces from the end of arguments, but other commands no longer will