consolidate alias handling
command alias handling refactored into a `mos_execAlias` function which accepts the system variable token for the alias, a pointer to the args string, an optional pointer to a “save token” to save a copy of the expanded alias, and an `in_mos` boolean.

`mos_runOrLoadFile` now uses this new function instead of doing its own alias handling.  load and run types now therefore support multi-line aliases

command depth detection has been reworked to use a global variable for the depth.  the depth is increased around calls to `mos_exec` inside the `mos_execAlias` command as well as the handlers for the `exec` and `obey` commands.  therefore MOS will now detect when there is an infinite loop of running scripts, as well as infinite loops resolving aliases