fix `run` command to be MOS 2 compatible
changes to how the `run` command interpreted its arguments introduced a bug that made it incompatible with commands written for MOS 1 and 2

the command `run . <filename>` should run the currently loaded program at the default address, passing in `<filename>` as the argument string.  instead the parser was seeing that `.` is not a number, and then passing the whole argument string

a side-effect of this is that `run <filename>` would work - but that behaviour is not desired, as ideally in the future that _should_ be equivalent to `runfile <filename>`

this update ensures that `run` now properly checks its arguments.  you must now either use no arguments, or provide a numeric address or a dot.

if the first argument is present but non-numeric an error will be reported - this differs from MOS 1 and 2, but is more correct behaviour, and is essentially compliant with existing documented examples