speed boosts
in `serial.asm` this changes a check inside the two `*_serial_TX` routines to check against the “TX hold register”.  this allows transmits to happen as soon as the serial system is no longer saying “hold”.  in benchmark testing this is is providing a very significant boost to our comms speed from MOS to the VDP, from around 640KBit/sec to 877KBit/sec.  as this is a measure of data bits sent, and our comms config sents 1 stop-bit per byte, this is getting very close to the maximum theoretical data rate

also changes the `FLASH_CTRL` configuration to use a value of `0x08`, which sets ROM/flash memory access to zero wait-states..  this boosts the perfomance of calling MOS APIs by around 17%
(I could not see how to adjust the `__FLASH_CTL_INIT_PARAM` setting - this does not appear to be a configuration directly exposed inside ZDS, and is not set in the source.  for simplicity a constant is being used in the source)

these changes were copied from a commit by @tomm on his experimental MOS 2.4 agondev port