add a mos_getfunction API (0x50)
this function returns a pointer to a function.  in general these should be functions that follow the C calling convention, and will be used to expose raw C functions

parameters are:
C = flags (must be zero for now)
B = function number

returns:
A = status (0, 19 or 20)
HL = pointer to function, or 0 if invalid

for now exposes the SD card C functions, and also `f_printf` which we can’t easily expose as a MOS-style API

flags are included for future use.  this may in the future allow a way to go beyond 256 functions, or if/when we introduce MOS modules control whether functions from modules will be returned or not

when called from Z80 mode, will return 20 (invalid command) as you can’t use these functions in Z80 mode.  (“invalid command” chosen over “not implemented” as the API _is_ implemented, it just can’t be used)

if the function number is out of range, or flags are not zero, returns 19 (invalid parameter)