some function and API tweaks
the addition of the `mos_getfunction` API means we need to make sure the functions we are exposing aren’t likely to change their arguments in any way in the future

`substituteArgs`’s last argument is now explicitly `BYTE flags` to match the `mos_substituteargs` API

`resolveRelativePath` now uses a pointer for its `length` argument instead of a direct int.  this will allow the function to change in the future to be able to calculate the length when the destination buffer pointer is null, as other path processing functions do.  the API implementation of `mos_getabsolutepath` which uses this function has been changed accordingly to store the provided length to the scratchpad, and restore that value before returning