fix for BASIC’s OPENIN/OPENOUT/OPENUP
BBC BASIC passes a control character (CR) terminated filepath to the `mos_fopen` API….

the documentation says filepaths should be zero terminated, however the reality is that fatfs interprets characters 0-31 (any control character) as a termination character, and so would tolerate/accept these filepaths

to allow `mos_fopen` to tolerate these filepaths, and thus fix the various versions of BASIC, the `resolvePath` function now works with control-character terminated strings.  this is done by finding the first control character in the path string and then making a copy of the string to that point

as a result of the changes to the `resolvePath` function, `mos_resolvepath` will now also be tolerant of control-character terminated paths

other `mos_*` APIs however generally won’t become fully tolerant of control-termination, and still require zero terminated paths strings.  these APIs tend to perform variable expansion on the input path first before passing the result of that on to another function to perform their action (which will call `resolvePath` with a flag to omit the variable expansion step).  variable expansion inherently can only work with zero terminated strings.  however, a variable expansion could introduce a control character into the path, and that will now be interpreted as a terminator by the underlying function

in general as the documentation has always stated that file paths should be zero terminated, this seems like an acceptable state for the API to be in.  the documentation will be updated to reflect this behaviour